requirements.txt
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
ghp-import==2.1.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-material>=5.3.2
mkdocs-material-extensions>=1.0
packaging==22.0
python-dateutil==2.8.2
PyYAML==6.0
pyyaml_env_tag==0.1
pymdown-extensions
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
watchdog==2.2.0
